+++ packages/lib/box/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.71](https://github.com/andremichelle/openDAW/compare/@opendaw/lib-box@0.0.70...@opendaw/lib-box@0.0.71) (2026-02-06)
+
+**Note:** Version bump only for package @opendaw/lib-box
+
 ## [0.0.70](https://github.com/andremichelle/openDAW/compare/@opendaw/lib-box@0.0.69...@opendaw/lib-box@0.0.70) (2026-02-03)
 
 **Note:** Version bump only for package @opendaw/lib-box