+++ packages/studio/enums/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.62](https://github.com/andremichelle/openDAW/compare/@opendaw/studio-enums@0.0.61...@opendaw/studio-enums@0.0.62) (2026-02-06)
+
+**Note:** Version bump only for package @opendaw/studio-enums
+
 ## [0.0.61](https://github.com/andremichelle/openDAW/compare/@opendaw/studio-enums@0.0.60...@opendaw/studio-enums@0.0.61) (2026-02-03)
 
 **Note:** Version bump only for package @opendaw/studio-enums